Split sidebar thread list derivation
