+++ docs/DECISIONS.md
@@ -278,3 +278,4 @@
 - Pull request command integration coverage lives in `WorkspacePullRequestIntegrationTests`, not `WorkspaceModelTests`. Remote PR view/checks/diff execution through SSH, slash `/pr` command dispatch, and PR command composer prefills cross workspace commands, slash parsing, SSH remote execution, tool cards, and composer state; keep pure GitHub CLI argument construction in tool tests and remote git builder tests.
 - Workspace activity surface derivation lives in `WorkspaceActivitySurfaceBuilder`, while section and item DTOs live in `WorkspaceActivitySectionSurface.swift`. `WorkspaceActivitySurface.swift` should remain the Codable root payload and delegate plan rows, recent event rows, tool/source/artifact summaries, final-answer snippets, handoff summary copy, and section assembly to the focused builder so native SwiftUI, static HTML, and Playwright continue sharing one activity contract.
 - Tool artifact state and preview derivation live in `QuillCodeToolArtifactSurface.swift`, not in `QuillCodeToolCardSurface.swift`. Tool cards own status, actions, density, review substates, and card-level artifact grouping; artifact kind detection, image/document preview metadata, href/detail labels, and bounded text preview extraction evolve beside the artifact DTO so transcript projection, activity, native cards, and static HTML share one artifact contract.
+- Sidebar thread list derivation lives in `QuillCodeSidebarThreadListBuilder.swift`, not in the aggregate `SidebarSurface` DTO. `SidebarSurface` keeps the stable public API consumed by native SwiftUI, static HTML, and search dialogs, while filtering, pinned/recent/archived partitioning, and date-bucket sectioning stay in the focused helper.
